requirements.txt
# local package
-e .

# external requirements
click~=8.0.4
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.4.1
numpy~=1.22.3
matplotlib~=3.5.1
scikit-learn~=1.0.2
tldextract~=3.2.0
kafka~=1.3.5
setuptools~=57.0.0
seaborn~=0.11.2
src~=0.1.0